requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
click==7.1.1
dash==1.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap-pagination==1.7.1
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-plotly-dash==1.1.5
django-timezone-field==3.1
dpd-components==0.1.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-Compress==1.4.0
future==0.18.2
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.7.0
plotly==4.6.0
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
regex==2020.2.20
retrying==1.3.3
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.1
whitenoise==5.0.1
XlsxWriter==1.2.8
